server/requirements.txt
certifi==2019.6.16
chardet==3.0.4
dataclasses==0.6
Django==2.2.3
django-cors-headers==3.0.2
django-mongodb-engine==0.6.0
djangorestframework==3.10.1
djangotoolbox==1.8.0
djongo==1.2.33
idna==2.8
pymongo==3.8.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.2.4
urllib3==1.25.3
watson-developer-cloud==2.10.1
websocket-client==0.48.0
